Added `UtilTrack::FindMatchedRoads()`, which find all roads that match the given track.  Modified SQVertexing so that it can be used outside the Fun4All macro.  Did a minor fix under `interface_main/` to suppress warnings.

+++ interface_main/SQCalHit_v1.cxx
@@ -15,8 +15,8 @@ ClassImp(SQCalHit_v1);
 
 SQCalHit_v1::SQCalHit_v1(): 
   _hit_id(std::numeric_limits<int>::max()),
-  _detector_id(std::numeric_limits<int>::max()),
-	_element_id(std::numeric_limits<int>::max())
+  _detector_id(std::numeric_limits<short>::max()),
+  _element_id(std::numeric_limits<short>::max())
 {
   _cells.clear();
 }
